chore: Rename service file.
